add error handling in user-ms

+++ microservices/users/src/routes/user.routes.ts
@@ -3,6 +3,7 @@ import { CreateUserDTO, UpdateUserDTO } from '../dto/user.dto';
 import UserController from '../controller/user';
 import createUserSchema from '../api/schema/user.schema';
 import passwordEncryption from '../utils/bcrypt';
+import { BadRequestError } from '../helpers/error-handler.helper';
 
 
 class UserRoutes {
@@ -29,38 +30,52 @@ class UserRoutes {
         router.post('/', async (req: Request, res: Response, next:NextFunction) => {
             const {error} =createUserSchema.validate(req.body);
             if (error) {
-                return res.status(400).send({success: false, error: error.details[0].message})
+                throw new BadRequestError(error.details[0].message)
+                // return res.status(400).send({success: false, error: error.details[0].message})
             }
             const hashedPassword = await passwordEncryption.hashedPassword(req.body.password)
             const payload:CreateUserDTO = Object.assign(req.body, {'password': hashedPassword});
-            console.log("payload", payload);
             
             const result = await controller.create(payload);
             return res.status(201).send(result)
         })
 
         router.get('/', async (req: Request, res: Response, next:NextFunction) => {
-            const result = await controller.getAll();
+            const result = await controller.getAll()
             return res.status(200).send(result)
         })
 
         router.get('/:id', async (req:Request, res:Response, next:NextFunction) => {
-            const id = Number(req.params.id)
-            const result = await controller.getById(id);
-            return res.status(200).send(result)
+            try {
+                if(!req.params.id) {
+                    throw new BadRequestError("Please pass the user id")
+                }
+                const id = Number(req.params.id)
+                const result = await controller.getById(id);
+                return res.status(200).send(result)
+            } catch (error) {
+                next(error)
+            }
+            
         })
 
         router.delete('/:id', async (req:Request, res:Response, next:NextFunction) => {
+            if(!req.params.id) {
+                throw new BadRequestError("Please pass the user id")
+            }
             const id = Number(req.params.id)
             const result = await controller.deleteById(id);
-            return res.status(204).send({success: result})
+            return res.status(200).send({message: "User deleted successfully"})
         })
 
         router.put('/:id', async (req:Request, res:Response, next:NextFunction) => {
+            if(!req.params.id) {
+                throw new BadRequestError("Please pass the user id")
+            }
             const id = Number(req.params.id)
             const payload: UpdateUserDTO = req.body
             const result = await controller.update(id, payload);
-            return res.status(201).send(result)
+            return res.status(200).send(result)
         })
